Fix some conversion errors cropping up in HEAD on OSX (new glew related?)

+++ RenderSystems/GL/src/OgreGLRenderSystem.cpp
@@ -487,7 +487,7 @@ namespace Ogre {
 			{
 				GLint buffers;
 				glGetIntegerv(GL_MAX_DRAW_BUFFERS_ARB, &buffers);
-				rsc->setNumMultiRenderTargets(std::min(buffers, OGRE_MAX_MULTIPLE_RENDER_TARGETS));
+				rsc->setNumMultiRenderTargets(std::min(buffers, (GLint)OGRE_MAX_MULTIPLE_RENDER_TARGETS));
 				rsc->setCapability(RSC_MRT_DIFFERENT_BIT_DEPTHS);
 				if(!GLEW_VERSION_2_0)
 				{
